tests: cleanup tests protypes and library
The main reason is in unification in test result checks.
And also this will simplify addition of prints in case of test failure.

+++ tests/ext_global_func.c
@@ -1,6 +1,6 @@
 #include "test_types.h"
 
-int ext_global_func(int type)
+long ext_global_func(int type)
 {
 	if (atoi("3") == 3)
 		return function_result(type);
