+++ tests/lib_global_func.c
@@ -1,6 +1,6 @@
 #include "test_types.h"
 
-int lib_global_func(int type)
+long lib_global_func(int type)
 {
 	return function_result(type);
 }
