+++ tests/lib_static_func.c
@@ -1,6 +1,6 @@
 #include "test_types.h"
 
-static int __attribute__ ((noinline)) lib_static_func(int type)
+static long __attribute__ ((noinline)) lib_static_func(int type)
 {
 	return function_result(type);
 }
